Main/requirements.txt
fastapi
uvicorn
firebase-admin
twilio
apscheduler
requests
python-dotenv
